Adding react-router in project
